Added custom style for "xs" select and the select itself to Navbar; moved providers to main.jsx to enable fetching inside App.jsx
