+++ sca-frontend/src/main.jsx
@@ -1,14 +1,32 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./App";
+import AuthProvider from "react-auth-kit";
+import createStore from "react-auth-kit/createStore";
+import { CssBaseline, ThemeProvider } from "@mui/material";
 
 import "@fontsource-variable/montserrat";
 import "overlayscrollbars/overlayscrollbars.css";
 import "./styles.css";
 import "./monaco/init.js";
 
+import App from "./App";
+import { theme } from "./theme";
+
+const store = createStore({
+    authName: '_auth',
+    authType: 'cookie',
+    cookieDomain: 'localhost',
+    cookieSecure: false,
+    cookieSameSite: 'lax',
+});
+
 createRoot(document.getElementById("root")).render(
     <StrictMode>
-        <App/>
+        <AuthProvider store={store}>
+            <ThemeProvider theme={theme} noSsr>
+                <CssBaseline enableColorScheme/>
+                <App/>
+            </ThemeProvider>
+        </AuthProvider>
     </StrictMode>,
 );
