+++ sca-frontend/src/utils/index.js
@@ -1,0 +1,2 @@
+export * from "./api";
+export * from "./fileTypes";
